[fidl][rust] Support DebugLog stub for Host Fidl Builds.
This change adds a stub for the encoding and decoding of the FidlDebugLog on host builds. This
stubbing allows FIDL libraries shared between host and fuchsia in
particular fuchsia-io2 to be used seamlessly as long as the unsupported
handle type isn't called from host. This is useful if you want to use a
small subsection of the library that is compatible across both host and
fuchsia but don't want to define more complex build rules. The build
will still fail if you attempt to use the DebugLog from host as
expected.

Bug: 39827
Change-Id: I9c210a7cb14cc06da775671a30a416f53167600c