[Brightness] Move request processing to a singleton class.
Essentially all the code has been moved from main.rs to control.rs. It is
untouched apart from adding self in front of required variables and a few
'?'s have been changed to handle the errors.

I have made several patches to show the transfer from one file to another.
Use patch sets 1 and 3 to see changes from main.rs to control.rs.

The use of several instances of the server task with a common
struct (Control) is favoured over merging streams because we need
to keep track of the streams for the implementation of hanging gets.

This CL has been waiting for so long that when I came to rebase it I
had to do a lot of changes with f64, hence all the backlight and other
extra changes in the last couple of patch sets.

I have spent two weeks trying to get an appropriate FIDL test, even
with the help of others have not managed it. I will add a client
test when I have a simple example to copy. My CLI client tests work.

Test: The tests still run.  Bothe unit and client test program.
Test: The display still works in auto and manual modes.
Bug: 36123
Change-Id: Ic668e75ea3cb687f51b7bc6e8e5626df6c42aa18