[nmh][network-manager] Bring interface.rs from netcfg.
Network manager will take over the function of netcfg regarding
interface discovery, oir management and network configuration.

interface.rs does the interface naming, keeping track of interface name
persistance. Bringing that file in so as to keep that functionality
unchanged.

No changes made to the code, other than those needed for it to compile.
All existing unit tests pass.
This code is not yet used. It will be used by the next CL.

Change-Id: I06e0f970358eadba5fe033f31273660be58c5fe0