[kernel] clear cmdline entropy from zbi.
This CL modifies the zbi parsing process by deleting the cmdline
arguments after they are processed. One of the arguments might include
entropy for the kernel cprng and we don't want it to be available to
other processes with access to the ZBI.

It looks like no other programs are using the CMDLINE zbi objects,
so this change should be safe. I also used a modified version of change
333450 to scan the memory and look for the entropy in any vmo, the only
one that showed up was the zbi vmo.

NOTE that this commit clears ALL the zbi objects of type CMDLINE and not
just the parts related to the kernel.entropy-mixin, because that might
appear in the middle of the payload, and appear multiple times with
different lengths. Instead, it's easier to just clear everything if it
is no longer needed.

The whole process is less than ideal, because the ZBI shouldn't be used
to pass secrets around, and there are no guarantees about what other
copies are lying around in memory. Sadly, this is the best we can do to
guarantee some entropy during early boot.

I added a test to the global_prng to make sure that the zbi does not
contain the entropy-mixin parameter. To run it, you can use: `k ut
global_prng`.

TEST=Tested in Astro, VIM2 and QEMU.

Change-Id: Iba72eaa90740d42d0309735dd4c6925ff7f28755