[fonts] Replace "base" font packages with local font files
In the font manifest generator and schema, eliminate the use of "base" vs
"universe" font packages, and replace them with "local" vs "download".

Local font files will be available to the font server through
`/config/data`. Downloadable font packages will be fetched through the
font_resolver as before, but the font server will have no inkling of
whether or not they are cached.

Corresponding changes to the font server to come in later CL.

Bug: 8904
Change-Id: I5bfde47ce8f745e484a923c0f694807bcb7263ef