[fidl][transformer] Add more transformer tests.
This CL adds many tests to the transformer from several disparate CLs:

* fxr/337061
* fxr/335800
* fxr/336588
* fxr/337309

Those CLs are in-flight and still need significant work. This CL is
designed to simply add the tests, so that they are visible to the entire
FIDL team and to unblock porting the tests to GIDL. Tests that do not
currently pass are commented-out.

This CL also makes a change to the C generator, which had a bug where
tables weren't being forward-declared properly. This affected the
transformer tests, since C headers were being generated from the
transformer tests. (Tables cannot be used in C, but other data
structures can reference them, so at least the type name must be made
available.) Thanks to yifeit@google.com for this change!

Test: fx set bringup.x64 --with-base //garnet/packages/tests:zircon && fx build && fx qemu -kN -c zircon.autorun.boot=/boot/bin/runtests+-t+fidl-test

Change-Id: I0cda8816a1170fe10e5c60c841ae15d83cdcfe5f