[fidl_codec] differ error messages
Currently we don't know if a message is a request or a response.
To select one of them, we try to decode a request and a response
and select the one which decodes. That means that the one which
doesn't decode can generate error messages which are not pertinent.

This patch records the error messages for both the request and the
response and displays the errors only if we haven't been able to
decode something correctly.

The messages are improved a little bit with the absolute offset in
the buffer where we have the error.

Previously, error_count wasn't set for all the errors.
Now it is set => fix xunion test which wrongly assert there was no
error.

Change-Id: I28dafe6dc656191e06b97fc15d17ddb6f0dc1126