[componentmgr] Disable excessive logging in Hub for used dir
The Hub logs out when a service is used multiple times and it
cannot add it to the 'used' dir. This was done because it was
not possible to distinguish between "node already exists" and
other possible errors from adding to a directory.

However, this causes a lot of spam since devcoordinator
frequently creates new connections to process.Launcher

Remove the logging from Hub for now.

Test: Ran on qemu. Logging does not appear.
Change-Id: Ib1800bd6854e972fcbae437f497d8f921cf0bad9