[kernel] Add constexpr ExtractBits.
There is an existing macro BITS that overlaps with this to an extent.
After being bit by a bug involving the order of arguments I think it
would be useful to have a version with c++ constexpr magic and
static_asserts to be more userfriendly.

Ideally we would drop the macro but it is used in some pretty important
kernel code that isn't well tested so I will defer on that for now.

This also modifies the valpow2 to support 64bit use cases but that is a
pretty minor change with few uses so it seems ok to just piggy-back on
this CL.

Test: Added unit tests.
Change-Id: Id7b6d425966906840568bdbc9f02f348b0f8a747