[firebase] Add HTTP request and response code.
This is stage 2 of a CL train to move the IDtoken->FirebaseIDToken
exchange from the AuthProvider into the FirebaseAuth library.

This code is taken from the GoogleAuthProvider, before the C++ version
was deleted in fxr/322438 with the following changes:
- Removed the now unused SetAuthorizationHeader and
SetUrlEncodedBodyMethods
- Added a trivial FirebaseRequestBuilder class to encapsulate the
constant strings that were previously running free in
GoogleAuthProvider.

Note that these classes are unit tested but not not yet used.

Change-Id: I815a761230e708de8780cbeea01e4548be7c1d0c