[gpt] Remove async bind thread
Until the DDK has better support for this type of operation, remove the
async bind thread. It races with unbind and fixing it seems unworkable
in the short term. This fixes a common test flake.

Convert gptpart_device_t to PartitionDevice. Added PartitionTable class
shared by PartitionDevices

Add FakeDDK-based test for GPT driver.
Most of GPT is tested in GPT ulib tests, and are outside of the scope of
these tests. This test verifies the driver binds and unbinds correctly.

Bugs: 34771
Tests: Added, blktest on mounted partitions, and manual testing of
booting with attached GPT disk.

Change-Id: Ide049e6c0cda83c1ad585b96f15d70a1391c1a5f