[bt][bt-gap] Add Generic Access Service for LE
Implement the Generic Access Service for LE. Right now, the HostDispatcher
spins up an async task which the Generic Access Service runs in when
HostDispatcher is initialized. This task stores a HostDispatcher which
it queries to respond to the Generic Access Service characteristics
value requests. Whenever HostDispatcher recognizes a new adapter, it
spins up a per-adapter GasProxy task which will forward incoming peer
GAS requests to the Generic Access Service through an async::mpsc::channel.
Right now the appearance variable is hardcoded in as Appearance::Display,
but this will need to be changed to be host-dependent.

Bug: 37351
Test: `$ fx run-test bt-gap-tests` will run the new tests for this module,
      which are specifically generic_access_service::tests::{
          test_change_name, test_get_appearance, test_gas_proxy,
	  test_invalid_request
      } which test the functionality of the GAS task and proxy, and  within
      host_dispatcher::tests::change_name_no_deadlock tests that a part
      of the code that used to deadlock no longer does.
Change-Id: I11df732a28e026a76f5de6cd123a0a4ba3fd5fa2