[pkgfs] remove all reliance on minfs
The dynamic index being stored on minfs is a major source of performance
challenges, and at this point represents a misdesign. It is removed without
immediate replacement, which has one major drawback:

Any ephemeral packages not used between last boot, and an OTA, will be
garbage collected by the OTA.

This drawback is actually somewhat of an advantage as of today, as the system
ABI is not stable, the retention of such packages is as yet, somewhat
dangerous - they should be checked for updates before being launched again.

A future design will restore some of the retention functionality by way of
the resolvers holding a cache of packages, and later the garbage collection
mechanism becoming lazy instead of aggressive.

To keep cache packages working, enable the cache_packages index added
in https://fuchsia-review.googlesource.com/c/fuchsia/+/332211 .

Bug: 37778
Change-Id: I1b49f0b5ccbccbe797b63b0ef95d3993d54e8603