[wlan][mlme] Implement BSS max idle
This causes a client to disassociate after the idle time is up. It will
not deauthenticate the client: the SME will do so with the association
timeout.

This also wraps RemoteClient::state in a StateMachine that is always
inhabited, because the ownership needs to be split from RemoteClient so
we can update the active timeout event ID while also borrowing
RemoteClient.

Test: Unit tests
Bug: 37891
Change-Id: Ibb0d4ce47933544c8c7705b58f2e2fd3ddb10425