[testrunner] Have testrunner optionally use runtests
This change starts respecting the `-use-runtests` switch (currently a
no-op). -use-runtsts is _only_ set now on profile builders (for coverage
reasons) - and the behaviour that this change introduces is that if this
switch is not passed, then run-test-component will be used instead.

The use of run-test-component over runtests will ultimately allow us to
start pulling test packages in ephmerally.

Note: bringup and perf builders are indeed still using runtests as they
use a different codepath that does not go through this tool.

Bug: 39738

Change-Id: Icbd8467d0ce45fe5e47f9e6bcfa2f93cc1ca5df7