[appmgr] Add sl4f component alias to allowlist
host driven OTA tests need to register a repository as "fuchsia.com" to
serve as the target build to be installed, but it also wants to start
sl4f using the build the target is currently running, so it registers
that repository using a different hostname.  This change adds that test
hostname to the deprecated shell allowlist.

Bug: 39652
Change-Id: I3db82f619cabba388b743dd35e0db3fe67956024