[componentmgr] Fixed OPEN_FLAG_DESCRIBE when opening directory-based framework capabilities
Currently if any process attempts to open a directory-based
framework capability (such as the hub), the flags passed in are
replaced with a hardcoded version (READ|WRITE).

However, problems arise when one of the incoming flags was
OPEN_FLAG_DESCRIBE. The client that initiated the open request
with this flag waits for an OnOpen event on the channel's
event stream but because the flags are replaced, the server never
sends the OnOpen event and the client blocks indefinitely.

This behavior was seen when attempting to open the hub from
within itself (by accessing exec/in/hub).

Fixed this by adding OPEN_FLAG_DESCRIBE to the hardcoded flags
for framework capabiltiies.

Updated hub_integration_test to send the DESCRIBE flag and
block on the DirectoryProxy's event stream until an OnOpen event
occurs. Without changing the hardcoded flags, some integration tests
would hang. After changing the flags, all tests pass.

Test: Ran all hub integration tests.
Change-Id: Iaf41508047d84bb134867ce5fc4b999e7e6005c0