[amlogic-video] Fix decoding after stream resize
This was causing the hardware decoder to hang (because it thought the
old reference frames were a weird size) and also caused video frames not
to be returned for reuse. We weren't catching this in the conformance
tests because they always start at the largest size and resize down, and
we don't reallocate frames in that case.

Testability: New additions to TestVP9.Decode to reallocate frames.

Change-Id: Iff5f1f0dcf162ae5b3ee400fe1f4f67e3def5afe