[kernel] split thread_t from Dispatcher
More exactly ThreadDispatcher. This continues the program
of modernizing the core codebase and continuing the work
of comment#1 of fxb/30771 and now tracked in 40105

This is the first step, which is splitting thread_t which
adds an additional heap alloc which is technically slower
but gets lost in the noise and the perfcompare bots do not
detect any significant regression: https://goto.google.com/uqsptb

Lifetime in this change stays the same.

This change does not introduce new behavior, it is a refactor
however a intermediate state now is tested via kernel unit tests
use 'k ut all' which the bots also run.

BUG = 40105

Change-Id: Ia270f2403d0097d5ecc84b76370332cccffced5f