[modular][xdr] Fix cloud provider bug
Modular config xdr parsing does some custom logic for parsing the
sessionmgr.cloud_provider and session_shell.display_usage fields because of
XDR's limitations with enums. The logic is meant to set the configs' FIDL
fields on XdrReads, but it was also doing so on XdrWrites.

This updates that logic to only occur on XdrReads.

TEST=
fx run-test modular_lib_tests -- --gtest_filter="ModularConfigReaderTest.GetConfigAsStringDoesntChangeValues"

Change-Id: Ic272cf0782bbf6bf0f915b1824d79361f76152c4