[linux_runner] Disable the ReuseExistingStatefulParition test case
While attempting to replace SafeStack with ShadowCallStack for arm, the
linux_runner_unittests test was failing from a segfault. We believe this
is due to a race condition between the tests, probably caused by some
gRPC internals.

The segfault occurs in `ReuseExistingStatefulParition` but can silenced
a number of ways:
- Disabling at least one of the other tests
- Avoid reading the `handle` at the end of `ConnectToStartupListener`
- Adding a log right after declaring the handle in
  `ConnectToStartupListener`
- Moving the SetUp and TearDown logic to the start and end of each test
  function
- Probably others to be discovered...

This will unblock the next Clang roll.

Bug: 40751, 27339, 39908
Change-Id: I33425ae9b3948b5c6747161628f055c855cbaa86