[scenic] Flatland Transform Graph
This CL externalizes the transform hierarchy into a separate helper class. Right now, all this CL does is move code around, but it prepares us for the next CL, which will create a TransformGraph factory with properties that aid in cross-instance links.

This CL also internalizes the std::multimap that manages the children lists, along with the method for generating cleanup code, into a single place. This logic will be useful for non-Flatland instances that still want to interact with the 2D graph, such as 2D-3D interop code, and display code.

Change-Id: Icae6a5719b238a95a71d8520b48ce22301a1a84b