[debugger] Describe call destinations.
Computes the absolute address of the destination of the most common call
instructions on both x64 and ARM. Attempts to resolve these as a symbol
and adds them as comments in the disassembly dump.

Call lines now look like this:

   0xea2fb7d8b8  call    0x3753     ➔ zx::object_base::reset(…)

Adds an option to the location formatting to omit the file/line
information which is used for the disassembly dump.

Convert the location formatting options to a structure since there are
now too many to practically pass individually.

Bug: 40996

Change-Id: Ic911d038fc0b688752e31a99a1d2792ad8ebbac5