[gidl] Add test for fuchsia.sys.Launcher/CreateComponent
This CL adds a test to transformer.gidl that mimics a CreateComponent
method request in fuchsia.sys.Launcher. This method revealed a bug in
the transformer, fixed in I39d267947604e0e3e5fae7e63588e87038ee64ce.
GIDL does not yet support handles, so the struct is not exactly the
same, but it is similar and encodes to the same bytes.

Test: run all GIDL tests

Change-Id: Icd3634e0e05057fadffbbda4391613b04e109ac9