[fidl][transformer] Testing: Fix primitive sizing
Tests for I157128e6dcd4dd9aca478b6e78e1185f8eb989f9

In the case of enums and bits, we have a coding table and need the
unaligned size when they are within structs. Otherwise, we skip too many
bytes, and end up reading past the |src_bytes| buffer.

Removing unused argument position to `InlineSize`. (We may need to
re-introduce this when we add diagnostics on failure, rather than
assert.)

Adding an implementation note on an imprecise use of `InlineSize`.

Test: fx qemu -kN, then runtests -t fidl-test
Change-Id: I17b41202df9f512e4534be711ec094db70093c0c