[virtualization] Refactor FakeNetstack
This refactor will eventually enable us to test guests with multiple
network devices. The major changes are to split out the device handling
code in the fake netstack so it can be reused, and to switch to a
promise based API. To differentiate between devices we now refer to them
by their MAC address.

Another notable change is that FakeNetstack now starts its own thread to
run a message loop that handles the FIFO waiters. This was done to
enable the asynchronous promise API.

Change-Id: Ib58c25a3f225595ecae179e9531a217b3d65e684