[camera][controller] Ensure proper shutdown of client
Without this change, when the client would get disconnected
we would begin the process of shutting down the stream pipeline.
In the process of shutting down, some of the member variables
of the parent nodes were getting destructed before the child nodes.

This was causing a problem where upon shutting down of the output
stream we would ask ISP to stop streaming, but the ISP stream protocol
pointer would be invalid. This change ensures that the child nodes
are always destructed completely before the parent node is.

Test: Added a test case to validate this.

Change-Id: Ic377aaa12221a5be3a071b7a855a609fb27626b8