[debugger] Sync Interface for Process Limbo Provider.
The ProcessLimbo fidl API is inherently async. So in order for the
debugger to call it sync (for IPC commands like status), it requires
wrapping the async calls with state tracking so that the agent is always
reflected with the current state.

Some callbacks to the provider will be added in a future CL.

TEST=Ported unit tests.
     fx run-test debug_agent_tests

Change-Id: Ia2c7c7df05390188ba2ac7babd9246b6feca0896