[mdns] fix sequencing error in publication
This CL fixes an error in which an InstanceResponder will attempt to
reannounce the instance before the responder is properly provisioned
to do so. This happens when the client implements
PublicationResponder and fires the Reannounce event before the
instance probe is complete.

TEST: The netemul test now fires the Reannounce event immediately
after calling PublishServiceInstance. The test now fails without
the fix and succeeds with it.

Change-Id: I25b83df95a653b595aef63971fb73f8d0a4a51af