[zircon][dynamic_loader] Introduce Dynamic Linker Break on Load.
This new property permits programs to set up a property on a process
handle that will make the dynamic linker issue a debug trap on the
following scenarios:

- On startup, after all dsos have been loaded.
- After startup, after every dso has been loaded.

The address of this debug trap is communicated through the new
`r_brk_on_load` member of the `r_debug` struct. The address of this
struct can be obtained through the ZX_PROP_PROCESS_DEBUG_ADDR property.

This will permits applications like the debugger to detect when a
process being debugged has loaded a new dynamic library (dlopen), by
comparing the address of the exception with the one on `r_brk_on_load`.

This DOES NOT change the existing behaviours of setting the magic
ZX_PROP_PROCESS_DEBUG_BREAK_ON_SET value to trigger the debug trap
on initial load, which is the current way zxdb gets the event that
the dynamic loader did it's job. Note that this magic value does
not handle subsequent loads (dlopen) as this approach does.

Bug: 40102

TEST=Added unit test.
     fx set <...> --with-base //garnet/packages/tests:zircon
     fx shell -- /system/test/sys/debugger-test

Change-Id: Iaa7d772372ea73f62adf17f78934e07d3a80c5c6