[fidl][tranformer] Fixes to table and array
Using the `dst_size` when tranforming array in order to properly pad if
needed.

When transforming envelopes with unknown types (which would happen with
wrong xunion ordinal for instance), the traversal result was not
updated.

Re-wrote table transformation to be simpler: we go over each envelope
one by one, and transform them. This avoids copying incorrect data
(because the num_bytes of the dst envelopes will differ), and bugs in
the out-of-line offset updates.

Test: added four regression cases
Change-Id: I17888a47d342a514b444213e17dcd414e3cbf68f