[wlan][mlme] assoc_req should use correct lenngth
When there are more than 8 rates, they will be split into 2 IEs. In this
case, another 2 bytes is needed for the extended supported rate ID and
len.

Also the input rsne already contained ID and len so frame_len should not
need to add another 2 bytes. This actually hid the bug for extended
supported rates because it provided extra buffer.

Bug: 41203

Test: Included.
Change-Id: Id013c3d00670f8bd1e23bc0dd4492c8258e7cf33