[kernel][exception] Zero-initialize zx_exception_info_t
Zero-initialize the zx_exception_info_t that's about to be sent over a
channel to ensure that its implicit padding does not expose memory
that should not be exposed.

Add a static_assert to verify that zx_exception_info_t is allowed to
be copied.

Add zx_exception_info_t to user_ptr's exception of list of
non-ABI-safe types allowed to be copied.  A future CL will remove the
implicit padding and remove the type from the exception list.

No behavior change.

Bug: 30802
Change-Id: I2794310eae35423b5cdc4457f49e6eb3d7769626