[asan] Decommit Partial Shadow Pages.
This CL fixes an issue with large allocations in asan by decommitting
partial pages as well as whole pages during `__sanitizer_fill_shadow`.

While tracking a memory leak in libfuzzer, we ended up finding that
large allocations in asan would take space in the shadow memory but part
of that memory was never returned back to the OS.

`__sanitizer_fill_shadow` is used to Poison/Unpoison memory regions,
keeping track of the state of each memory byte in the asan shadow bitmap.

The asan shadow in fuchsia covers ~20TB of address space, and each byte
in the shadow memory corresponds to 1 << ASAN_SHADOW_SHIFT bytes (8
bytes). Initially, it is all initialized to zeroes, and does not consume
any memory as the zircon kernel does not allocate anything for it yet.

After the first read/write on each page, the kernel will map a
copy-on-write zero page, so if asan tries to write to that page, a new
page will be allocated.

These pages are part of a vmo (the shadow vmo), which are only returned
to the OS during `__sanitizer_fill_shadow`, if the user is trying to
unpoison a large amount of memory (larger than the threshold parameter).
However, the function only decommits whole pages, leaving partial writes
with just zeroes, but committed to memory.

This behavior is similar to what other operating systems do. However,
they have other mechanisms to free those resources. In Linux, for
example, there's a daemon that runs in the background de-duplicating
pages, removing the zero pages periodically. Also, in linux, most of the
big allocations end up being reused, so the asan shadow ends up writing
over and over the same memory ranges.

In Fuchsia, all these allocations end up being scattered over the
address space, leaving 1 or 2 pages lying around every time a new big
allocation is unpoisoned.

The fix I am proposing is to check in `__sanitizer_fill_shadow` whether
the edge pages (partial pages to be unpoisoned) end up containing all
zeroes, and if so, also decommit those pages.

Note that this only affects allocations that are bigger than 1 shadow
page (8 real pages ~32KB), and are also greater than the asan threshold.

Smaller allocations (<1 shadow page) will live forever, regardless of
the threshold. To fix that, we should also add a similar check on the
other branch of the main if condition in `__sanitizer_fill_shadow`

I added a test that tries to show this behavior. I tried to do it as close
as possible to "asan" code as the other test in the file, however, a
simple malloc(something_big)/free in a loop would be enough to reproduce
the behavior.

To compile it, run:

```
$ fx set core.x64 \
  --with-base //bundles:tools \
  --with-base garnet/tests/zircon:zircon_tests \
  --variant=asan/sanitizer-test \
  --variant=asan/zircon_tests
$ fx build
```

And you run them with:
`$ runtests -t sanitizer-test`

BUG=34570

Change-Id: I6ed6130ba8ae89f7005bb05510ae436fffe8c2de