[ldso] Redirect __asan_memset to internal version
Attempting to land this for Roland since it got reverted while
attempting a previous GI roll.

This should have been done before and lacking it was a fragile situation.
But it became an active bug when rolling to a toolchain build where the
ASan runtime DSO uses -fsanitize=shadow-call-stack (i.e. it's default in
the --target=aarch64-fuchsia configuration).

Bug: 27339
Change-Id: Ib19dd89582fd8082a5c62a817a143bd2900cd960