[FDF] Support configuring auto suspend
At times drivers want to auto suspend and resume by themselves based on their
idleness. This is particularly useful when hardware has proper support for
doing this(For ex: CPU device). This change adds support for system
power manager to configure auto suspend with the deepest sleep state,
power manager wants this device to go into when this device is suspending
automatically. Selective device suspend will return ZX_ERR_NOT_SUPPORTED when the
device is configured for auto suspend and will not call the device's
suspend hook, although the hook is implemented by the driver. On the other hand,
the System Suspend will ignore the autosuspend config and proceed with calling the suspend
hook of the driver if implemented.

Test: runtests -t ddk-power-test

Bug: 39905

Change-Id: I0539af36b0250a75fa8034c49d8d925b294afc58