[pseudo-fs-mt] {Token,Inode}Registry in ExecutionScope
ExecutionScope seems like a natural place to put "runtime related"
configuration of the tree nodes.  This way a specific client can decide
what is the scope of the renaming operation, as well as the scope of the
"inode" registry.  Or provide custom versions of those services.

Similarly, a mechanism to control new node construction will also be
scoped using ExecutionScope objects.

Change-Id: I097e8aba09ac12d904e354cacea58da33a0760ca