[modular] Don't restart session_agents on shutdown
Flake in stories_share_session_runners_test.

There are also multiple attempts to restart the FakeSessionAgent,
initiated when the FakeSessionAgent dies because some of its service
dependencies are shutting down. The UserIntelligenceProvider now avoids
restarting session agents if sessionmgr is terminating.

It's unclear if or how this bug contributes to the flake (which I
could not yet reproduce), but if the flake does recur, the issue can be
eliminated as a suspected cause.

Bug: 41150
Change-Id: Iea43ac87fde2d743eb7ae767b9d32752e678b078