[modular] Fix test manifest related to test flake
Flake in stories_share_session_runners_test.

There are errors related to a bad module manifest. The error does not
appear to break the test, but it is clearly logged. Manifest is fixed,
and I added additional testing for the kind of error in
module_manifest_xdr_unittest.cc. The new test won't prevent errors
elsewhere, but serves as a better example of what is not a valid
manifest.

It's unclear if or how this bug contributes to the flake (which I
could not yet reproduce), but if the flake does recur, the issue can be
eliminated as a suspected cause.

Bug: 41150
Change-Id: I695908f2adcc658ff98bb0c612e8c63377ff21ac