[zircon][display] Add a vsync watchdog.
A number of bugs in vsync handling have been revealed through test
flakes. In order to capture crash reports for more of these and help
us track them down, this change introduces a watchdog timer that will
crash the devhost if no vsync is delivered to a client for over a
minute.

Test: Build and run terminal under AEMU and on Astro
Change-Id: I46600907fa1f53d09ade67b7c37897d4c4341f7d