[amlogic-video][vp9] Disable compressed frame output
Compressed frame output isn't necessary unless we're doing 10bit video
(so that the reference frames are correct) or greater than 1080p (for
memory bandwidth, probably). If we disable it we can reduce memory usage
substantially.

One complication is that we have to keep around some of the old
VideoFrames after resize, since they may still be used as reference
frames. Therefore we can only get rid of VideoFramess after they're only
in use by the client (or not at all).

Fixed: 40062
Test: VP9 conformance tests on sherlock (plus integration tests).

Change-Id: I98e075c17124f41e2f78f699654a4816e424a739