[fs-management] break up mount.h into more granular headers
Break out some operations in mount.h that aren't mount-specific, or are
useful for the upcoming new admin functions, so that clients that are
using the new functions don't have to include all the mount functions
too. These new header files are transitively included by mount.h, so
there is no change in how clients import things for now.

Test: /system/test/fs/fs-management-test
Change-Id: I0f9c018bbc6a12847bf29a9b36d7b10f8e16dddc