[Fuchsia DevTools] gather thread CPU more often
This CL adds a new system to gather task structure and thread CPU data.
The task structure is gathered at a much lower rate than the per thread CPU
data to lower the overhead (amortize the cost). The per thread CPU data is then
gathered at a rate on par with the overall system CPU metrics (currently 10Hz).
This system is being done in addition to the current system as part of either:
a soft transition or a short term workaround (depending on how things play out).

fxb/41126

Change-Id: I67eda11b5b701d1b304f7718fd6756ebff9828b7