[kernel][vm] Add ZX_VMO_OP_ZERO
Adds a new VMO range operation ZX_VMO_OP_ZERO that performs efficient
zeroing of a range of memory. This operation is efficient as it uses,
where possible, markers in the page_list_ to indicate a page has been
zeroed, or is now zero, instead of actually zeroing.

Due to aspects of how bi-directional CoW clones are presently
implemented zeroing may cause additional pages to be committed if a
page has to be forked from a distant parent before a marker can be
placed.

Test: Added additional core tests.

Bug: 36474

Change-Id: I63e82fd631f2fe7c7c513fb9c3d667fcbd419297