[kazoo] Delete some now-unnecessary \n in outputs
A variety of the backends were outputting extra \ns to match the output
of abigen, but this is no longer necessary, so remove them. Some could
still be slightly prettier (often, they still end in an unnecessary
blank line), but it'd generally involve a check for "on last element?"
which doesn't seem worthwhile.

Bug: 33668
Bug: 39918
Change-Id: Id2f244b4240e9797858e6029e7a189685d0b9a05