[syscalls] Gate zx_debug_read and zx_debug_write
When kernel.enable-debugging-syscalls=false, make zx_debug_read() and
zx_debug_write() fast fail.

The implementation is trivial, however, testing was more complex as the
existing integration tests to test kernel arguments use the serial port
as the method of interactively running commands and querying for output.

This change adds a non-interactive method of running a command on the
target and capturing its output by using a smattering of qemu and kernel
command line flag to mount a host disk, run the desired command, and
capture the output back to the host disk.

Test: disable_syscalls_test refactored and updated
Bug: 38409
Change-Id: I7c09a197b8155b62401a8ff198fa4b564ba0d935