[acpi][battery] Minimal C++ migration
This does an intentionally minimal C++ conversion of the dev-battery
ACPI device. Changes not strictly required for conversion is test code
no longer #include the implementation, but rather uses a shared header
for definitions.

Motivation for this is general cleanup of the x86 board driver, which
will be easier if everything is C++.

Test: runtests -t x86-battery-test

Change-Id: Iaef6fc7a1d1e5f134a34cfa8b53df9c77990ec4b