[sdhci] Add support for 64-bit ADMA2 descriptor
The SDHCI spec defines a 96-bit descriptor format that supports 64-bit
system addresses and a 64-bit descriptor format that supports 32-bit
system addresses. Currently the SDHCI driver only uses the 96-bit
descriptor, but the AS370 controller only supports the 64-bit
descriptor. This change adds support for the 64-bit descriptor to enable
DMA on AS370.

The following changes were also made to the SDMMC protocol to decouple
it from SDHCI:
- Removed the SIXTY_FOUR_BIT host capability
- Renamed the ADMA2 host capability to DMA

Bug: 39697
Test: runtests -t sdhci-test
Test: Paved Eve, ran iochk
Test: SDIO initializes on AS370
Change-Id: I4e3df3e56b78a86a5ba6df64bed534a99fc42e85