[componentmgr] Bind to child also binds to parent
This change ensures that a request to bind to a child automatically
binds to the parent. Every child has an implicit dependency on its
parent. The parent is responsible for managing its children, may
dynamically provision them, and may want to monitor them. Thus, running
a parent along with its child ensures that the child won't encounter
unexpected conditions or behave incorrectly on account of its parent not
running.

This change is only a small step in the right direction. While it
guarantees that a parent will be started before its child, it will not
do the inverse of stopping the parent when the child stops (and there
are no other bindings). This is because component_manager doesn't yet
have a representation of a "binding"; further progress in this
direction will require adding this.

This change also moves eager bindings into the bind() call. This removes
the burden on call sites to bind to the eager children themselves.

Change-Id: I127a0fb3cddb503b41fd20028d1553624955a4af