[settings] Reland removing DEFAULT in audio source
Changes persistent storage to use the default value if it is unable to parse the stored value
that is requested. Because of the change to the AudioSettingSource fidl api, any DEFAULT enum
values stored would cause a crash. These changes are in device_storage.rs.

All other changes are from the original CL.
(https://fuchsia-review.googlesource.com/c/fuchsia/+/336009)

Tests: Wrote a test in device_storage to verify that if an invalid value exists in device_storage,
it instead returns the default value when requested.

fx run-test setui_service_tests -- --test

Change-Id: Ide3cb636f6afa82bf6e096eadae6237f4c6393f2