[inspect][rust] Get rid of "mut Inspector/Node"  to continue supporting component::inspector()
Without this we would need to force client structs to be borrowed
as mutable if they wish to record a property to a node.

With the previous approach we would have required something like
`component::inspector_mut`. But that returns a value wrapped in a
lazy_static, so it cannot be borrowed mutable and needs to be wrapped in
a Mutex.

We already had a Mutex guard on the Node.values. So this CL
gets rid of needing `&mut self` on `record_*` methods.

TESTED=fx run-test fuchsia_inspect_tests

Change-Id: I4eb290c8ae0f313d262f818a9a86cb84162a12e9