[amlogic-video] Allow using parser to copy data
For VP9 decrypted frames are in vdec memory (or in non-protected
memory), but the decoder can only read from regular protected memory. We
need to be able to use the parser to copy data, but without hooking it
up directly to the decoder (which would be incompatible with
multi-instance decoding). Add support for setting up a non-stream output
buffer and writing into that.

Testability: new Parser tests

Change-Id: I8de03b8e2cb837945333abfbddeef0c21015c239