[kernel][x86] CpuId: Add support for Leaf 6h, add definitions for HWP/EPB
fxr/336149 enables HWP (Hardware P-state selection) by default and uses the
IA32_ENERGY_PERF_BIAS register to guide the Energy/Performance Preference.
Add definitions of CPUID Leaf 6h, HWP, and EPB to the new CpuId code, so that
we can convert power/performance logic to use it and write tests for it.

Bug: 39391 Regressions of 130-230% on put_entry, update_entry, stories_simulation, get_page

Change-Id: I22509cc6ee592ceafd0e10f1b1e4c0c4256c3fe5