[debugger] LimboProvider provides a callback for new proceses.
This permits the agent to detect when new processes enter the limbo and
be able to send a notification back to the client. Current the client
works on manually polling, but sending a notifications provides for a
better UX.

Added handlers to be tracked over by the ProcessLimboManager (over on
the ExceptionBroker). Without that, connections would never get the fidl
events.

TEST=Added unit test.
     fx run-test debug_agent_tests

Change-Id: I3ab3d85c46a6867b22abe811d284d0862a7f52f7