[scenic] implement Present2 and OnFramePresented
This change adds another queue through the frame scheduler for piping
|Present2Info|s, used wherever existing Present1Callbacks are. These
Present2Infos give the frame scheduler the necessary information to
fire OnFramePresented events for Present2() calls, which it now does.
Thus, Present2() and OnFramePresented() are fully functional.

Some implementation details to note:

- We also add a ScheduleUpdateForPresent2() method, to be used wherever
ScheduleUpdateForPresent() is. Their common functionality is factored
out.

- We also modify a Session::Update to include both Present and Present2
support. We do this because Session::ApplyScheduledUpdates is the only
thing that needs to know about them, and we want to not have duplicated
logic.

Test: Converted Present1 tests over to use Present2. Added
Present2-specific tests in session_unittest.cc and
default_frame_scheduler_unittest.cc

Bug: 24605

Change-Id: I0f5b0f8435e70b7c9c6fecaca487ee5eea55101b